+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.0] - 2025-12-16
+
+- Comando /mascota para invocar una mascota del canal mascotas.
+- Posibilidad de controlar a Hepatitis B(ot).
+
+
 ## [1.0.0] - 2025-12-15
 
 - Lanzamiento inicial.